Move `InteractiveAsker` to `pip_review.py` and mark it as private
